+++ package.json
@@ -15,7 +15,6 @@
     "@tokilabs/lang": "1.0.3",
     "guid-typescript": "1.0.9",
     "reflect-metadata": "0.1.13",
-    "rxjs": "7.4.0",
     "tslib": "2.0.0",
     "zone.js": "0.11.4"
   },
@@ -25,20 +24,18 @@
     "@nrwl/cli": "13.4.5",
     "@nrwl/cypress": "13.4.5",
     "@nrwl/eslint-plugin-nx": "13.4.5",
-    "@nrwl/express": "^13.5.2",
     "@nrwl/jest": "13.4.5",
     "@nrwl/linter": "13.4.5",
-    "@nrwl/nest": "13.4.5",
+    "@nrwl/nest": "13.5.2",
     "@nrwl/node": "13.4.5",
     "@nrwl/tao": "13.4.5",
     "@nrwl/workspace": "13.4.5",
-    "@types/body-parser": "^1.19.2",
+    "@types/body-parser": "1.19.2",
     "@types/jest": "27.0.2",
     "@types/node": "14.14.33",
     "@typescript-eslint/eslint-plugin": "5.3.0",
     "@typescript-eslint/parser": "5.3.0",
     "autoprefixer": "10.4.0",
-    "@types/nodemailer": "^6.4.4",
     "cypress": "9.1.0",
     "eslint": "8.2.0",
     "eslint-config-prettier": "8.1.0",
